Removed vec3_t union. Access now via [] operator.
I think this makes the code now more readable.
Also in debug mode a range check could be implemented.
glVertexfv() can now get a pointer with the new GetPointer() function.

+++ src/Stream.cpp
@@ -297,7 +297,7 @@ void CStream::WriteStream(const CStream& stream)
         WriteBytes(stream.m_buffer, stream.GetBytesWritten());
 }
 
-size_t CStream::StringSize(const std::string& value)
+unsigned int CStream::StringSize(const std::string& value)
 {
     if(value.length() >= USHRT_MAX)
     {
